'addSendMessage'

+++ src/main/java/com/cqjtu/bysj/mapper/ChapterMapper.java
@@ -13,7 +13,7 @@ public interface ChapterMapper {
 
     void deleteChapter(@Param("chapterId") Long chapterId);
 
-    void deleteChapterByCourseId(@Param("courseId") Integer courseId);
+    void deleteChapterByCourseId(@Param("courseId") Long courseId);
 
     void updateChapter(Chapter chapter);
 }
