+++ src/main/java/com/cqjtu/bysj/service/CourseService.java
@@ -14,6 +14,6 @@ public interface CourseService {
 
     void updateCourse(Course course);
 
-    void deleteCourse(Integer courseId);
+    void deleteCourse(Long courseId);
 
 }
